+++ CHANGELOG.rst
@@ -18,7 +18,7 @@ Changed
 
 - Updated dataset item identifier from file content sha1sum to relative file
   path sha1sum
-
+- Changed manifest item storage from list ("file_list") to dictionary ("items")
 
 Deprecated
 ^^^^^^^^^^